chore(ui): reduce code duplication around diff detection
