+++ internal/ui/chat/generic.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	"github.com/charmbracelet/crush/internal/message"
-	"github.com/charmbracelet/crush/internal/stringext"
 	"github.com/charmbracelet/crush/internal/ui/styles"
 )
 
@@ -32,7 +31,7 @@ type GenericToolRenderContext struct{}
 // RenderTool implements the [ToolRenderer] interface.
 func (g *GenericToolRenderContext) RenderTool(sty *styles.Styles, width int, opts *ToolRenderOpts) string {
 	cappedWidth := cappedMessageWidth(width)
-	name := genericPrettyName(opts.ToolCall.Name)
+	name := humanizedToolName(opts.ToolCall.Name)
 
 	if opts.IsPending() {
 		return pendingTool(sty, name, opts.Anim, opts.Compact)
@@ -64,37 +63,11 @@ func (g *GenericToolRenderContext) RenderTool(sty *styles.Styles, width int, opt
 
 	bodyWidth := cappedWidth - toolBodyLeftPaddingTotal
 
-	// Handle image data.
 	if opts.Result.Data != "" && strings.HasPrefix(opts.Result.MIMEType, "image/") {
 		body := sty.Tool.Body.Render(toolOutputImageContent(sty, opts.Result.Data, opts.Result.MIMEType))
 		return joinToolParts(header, body)
 	}
 
-	// Try to parse result as JSON for pretty display.
-	var result json.RawMessage
-	var body string
-	if err := json.Unmarshal([]byte(opts.Result.Content), &result); err == nil {
-		prettyResult, err := json.MarshalIndent(result, "", "  ")
-		if err == nil {
-			body = sty.Tool.Body.Render(toolOutputCodeContent(sty, "result.json", string(prettyResult), 0, bodyWidth, opts.ExpandedContent))
-		} else {
-			body = sty.Tool.Body.Render(toolOutputPlainContent(sty, opts.Result.Content, bodyWidth, opts.ExpandedContent))
-		}
-	} else if looksLikeDiff(opts.Result.Content) {
-		body = toolOutputDiffContentFromUnified(sty, opts.Result.Content, cappedWidth, opts.ExpandedContent)
-	} else if looksLikeMarkdown(opts.Result.Content) {
-		body = sty.Tool.Body.Render(toolOutputCodeContent(sty, "result.md", opts.Result.Content, 0, bodyWidth, opts.ExpandedContent))
-	} else {
-		body = sty.Tool.Body.Render(toolOutputPlainContent(sty, opts.Result.Content, bodyWidth, opts.ExpandedContent))
-	}
-
+	body := renderToolResultTextContent(sty, opts.Result.Content, toolResultContentWidths{Body: bodyWidth, Diff: cappedWidth}, opts.ExpandedContent)
 	return joinToolParts(header, body)
-}
-
-// genericPrettyName converts a snake_case or kebab-case tool name to a
-// human-readable title case name.
-func genericPrettyName(name string) string {
-	name = strings.ReplaceAll(name, "_", " ")
-	name = strings.ReplaceAll(name, "-", " ")
-	return stringext.Capitalize(name)
 }
